refactor: move User types
