+++ src/app/(insights)/endorsements/_types/skill.ts
@@ -1,4 +1,4 @@
-import type { User } from './user'
+import type { User } from '@/types/user'
 
 export type Skill = {
   id: string
