+++ src/types/user.ts
@@ -1,5 +1,6 @@
 export type User = {
   id: string
   name: string
+  email?: string
   image: string
 }
